+++ src/components/StatusBar/styles.js
@@ -1,7 +1,5 @@
 import { StyleSheet } from 'react-native';
 
-import styleConstants from '../../styleConstants';
-
 const styles = StyleSheet.create({
   container: {
     position: 'absolute',