+++ src/components/StatusBarComponent/index.js
@@ -1,3 +1,0 @@
-import StatusBarComponent from './StatusBarComponent';
-
-export default StatusBarComponent;